#664 rename broadcast composable
